lab5/requirements.txt
praw
mysql-connector-python
beautifulsoup4
nltk
pytesseract
Pillow
requests
gensim
scikit-learn
numpy
pandas
matplotlib
seaborn
wordcloud